requirement.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
Django==2.0
django-tinymce==2.7.0
idna==2.7
lxml==4.2.5
Pillow==5.3.0
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.7
qiniu==7.2.2
requests==2.20.1
six==1.11.0
urllib3==1.24.1
